Initial move to Unity

+++ FastNoise/Vector2.cs
@@ -6,8 +6,8 @@ namespace FastNoise
 {
     public struct Vector2
     {
-        public readonly double x, y;
-        public Vector2(double x, double y)
+        public readonly float x, y;
+        public Vector2(float x, float y)
         {
             this.x = x;
             this.y = y;
